fix(metrics): remove residual kube provider surface
