+++ metrics/docs/adr/0015-interactive-quota-scope-and-kube-provider-role.md
@@ -6,13 +6,14 @@ Proposed (M5 — not yet implemented)
 
 ## Context
 
-The `kube` provider package exists for typed configuration before M5. Platform
-metrics must remain Kueue-backed; quota reads come from live Pod state.
+No `kube` provider package or configuration exists before M5. Platform metrics
+must remain Kueue-backed; a future quota implementation may read live Pod
+state.
 
 ## Decision
 
-- **`sources.quotas.interactive: kube`** binds interactive quota to the kube
-  provider—not to `sources.platform`.
+- M5 may add **`sources.quotas.interactive: kube`** only with a complete kube
+  quota provider; it must not alter `sources.platform`.
 - Scope id: **`quotas.interactive`** (`MetricScope.INTERACTIVE_QUOTA`).
 - Provider method: complete **`interactive_quota(user)`** model; no platform
   aggregation in kube.
@@ -21,8 +22,8 @@ metrics must remain Kueue-backed; quota reads come from live Pod state.
 
 - RBAC must allow **list/get Pods** in configured namespaces for quota, separate
   from Kueue ClusterQueue access for platform metrics.
-- Helm values must configure both Kueue platform source and kube quota source
-  when interactive quota is enabled.
+- M5 Helm values must configure both Kueue platform source and kube quota
+  source when interactive quota ships.
 
 ## References
